Create new component for dependent fields
