+++ src/app/shared/components/dependent-fields/dependent-fields.component.html
@@ -1,0 +1,24 @@
+<div *ngIf="dependentFields?.length">
+  <ng-container formArrayName="dependent_fields">
+    <div *ngFor="let dependentField of dependentFields" class="dependent-fields--internal-block">
+      <div [formGroup]="dependentField.control">
+        <app-dependent-field
+          [label]="dependentField.field"
+          [placeholder]="dependentField.placeholder"
+          [mandatory]="dependentField.mandatory"
+          [valid]="!(dependentField?.control?.touched && !dependentField?.control?.valid)"
+          [fieldId]="dependentField.id"
+          [parentFieldId]="dependentField.parentFieldId"
+          [parentFieldValue]="dependentField.parentFieldValue"
+          formControlName="value"
+        >
+        </app-dependent-field>
+      </div>
+    </div>
+  </ng-container>
+</div>
+
+<div class="dependent-fields--primary-block" *ngIf="isDependentFieldLoading">
+  <ion-skeleton-text class="dependent-fields__skeleton-loader--title" [animated]="true"></ion-skeleton-text>
+  <ion-skeleton-text class="dependent-fields__skeleton-loader" [animated]="true"></ion-skeleton-text>
+</div>
