+++ src/app/shared/components/dependent-fields/dependent-fields.component.spec.ts
@@ -1,0 +1,24 @@
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+import { IonicModule } from '@ionic/angular';
+
+import { DependentFieldsComponent } from './dependent-fields.component';
+
+describe('DependentFieldsComponent', () => {
+  let component: DependentFieldsComponent;
+  let fixture: ComponentFixture<DependentFieldsComponent>;
+
+  beforeEach(waitForAsync(() => {
+    TestBed.configureTestingModule({
+      declarations: [DependentFieldsComponent],
+      imports: [IonicModule.forRoot()],
+    }).compileComponents();
+
+    fixture = TestBed.createComponent(DependentFieldsComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  }));
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
